Add geotrellis.spark.util.SerializableConfiguration and fix HBase connection again
